Chore/mutation refactor (#194)
* chore!: refactor deleteGfDraftEntry

* chore!: refactor deleteGfEntry

* dev: abstract get_from_id()

* chore!: refactor updateGfDraftEntry (sans values)

* dev: use context loader instead of DataSource

* fix: pass correct user_agent to model

* chore: move entry meta values to their own input and add missing keys.

* chore: refactor submitGfDraftEntry

* chore!: change params for get_resume_url(

* feat: add dateCreatedGmt

* fix: sanitize input values

* chore!: refactor SubmitForm mutation (sans field values)

* fix: revert codeception package versions to 1.x

* chore: fix broken github workflow

+++ src/Mutation/DeleteDraftEntry.php
@@ -10,10 +10,17 @@
 
 namespace WPGraphQL\GF\Mutation;
 
+use GFCommon;
 use GFFormsModel;
 use GraphQL\Error\UserError;
 use GraphQL\Type\Definition\ResolveInfo;
+use GraphQLRelay\Relay;
 use WPGraphQL\AppContext;
+use WPGraphQL\GF\Data\Loader\DraftEntriesLoader;
+use WPGraphQL\GF\Model\DraftEntry as ModelDraftEntry;
+use WPGraphQL\GF\Type\Enum\DraftEntryIdTypeEnum;
+use WPGraphQL\GF\Type\WPObject\Entry\DraftEntry;
+use WPGraphQL\GF\Utils\GFUtils;
 
 /**
  * Class - DeleteDraftEntry
@@ -32,9 +39,13 @@ class DeleteDraftEntry extends AbstractMutation {
 	 */
 	public static function get_input_fields() : array {
 		return [
-			'resumeToken' => [
-				'type'        => [ 'non_null' => 'String' ],
-				'description' => __( 'Resume token of the draft to delete.', 'wp-graphql-gravity-forms' ),
+			'id'     => [
+				'type'        => [ 'non_null' => 'ID' ],
+				'description' => __( 'Either the global ID of the draft entry, or its resume token', 'wp-graphql-gravity-forms' ),
+			],
+			'idType' => [
+				'type'        => DraftEntryIdTypeEnum::$type,
+				'description' => __( 'The ID type for the draft entry. Defaults to `ID` ', 'wp-graphql-gravity-forms' ),
 			],
 		];
 	}
@@ -44,9 +55,18 @@ class DeleteDraftEntry extends AbstractMutation {
 	 */
 	public static function get_output_fields() : array {
 		return [
-			'resumeToken' => [
-				'type'        => 'String',
-				'description' => __( 'Resume token of the draft that was deleted.', 'wp-graphql-gravity-forms' ),
+			'deletedId'  => [
+				'type'        => 'ID',
+				'description' => __( 'The global ID of the draft entry that was deleted.', 'wp-graphql-gravity-forms' ),
+				'resolve'     => function( $payload ) {
+					$deleted = (object) $payload['deletedEntry'];
+					return ! empty( $deleted->id ) ? $deleted->id : null;
+				},
+			],
+			'draftEntry' => [
+				'type'        => DraftEntry::$type,
+				'description' => __( 'The draft entry object before it was deleted', 'wp-graphql-gravity-forms' ),
+				'resolve'     => fn( $payload ) => $payload['deletedEntry'] ?? null,
 			],
 		];
 	}
@@ -56,30 +76,24 @@ class DeleteDraftEntry extends AbstractMutation {
 	 */
 	public static function mutate_and_get_payload() : callable {
 		return function( $input, AppContext $context, ResolveInfo $info ) : array {
-			static::check_required_inputs( $input );
+			if ( ! GFCommon::current_user_can_any( 'gravityforms_delete_entries' ) ) {
+				throw new UserError( __( 'Sorry, you are not allowed to delete entries', 'wp-graphql-gravity-forms' ) );
+			}
 
-			$resume_token = sanitize_text_field( $input['resumeToken'] );
-			$result       = GFFormsModel::delete_draft_submission( $resume_token );
+			$id_type      = isset( $input['idType'] ) ? $input['idType'] : 'global_id';
+			$resume_token = self::get_resume_token_from_id( $input['id'], $id_type );
+
+			$entry_before_delete = GFUtils::get_draft_entry( $resume_token );
+
+			$result = GFFormsModel::delete_draft_submission( $resume_token );
 
 			if ( ! $result ) {
 				throw new UserError( __( 'An error occurred while trying to delete the draft entry.', 'wp-graphql-gravity-forms' ) );
 			}
 
 			return [
-				'resumeToken' => $resume_token,
+				'deletedEntry' => new ModelDraftEntry( $entry_before_delete, $resume_token ),
 			];
 		};
-	}
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * @throws UserError .
-	 */
-	protected static function check_required_inputs( $input ) : void {
-		parent::check_required_inputs( $input );
-		if ( ! isset( $input['resumeToken'] ) ) {
-				throw new UserError( __( 'Mutation not processed. The resumeToken must be set.', 'wp-graphql-gravity-forms' ) );
-		}
 	}
 }
